installers/polling_ui/requirements.txt
Pillow
StreamDeck
hidapi
